Rename folders to use singular nouns only
