+++ backend/src/helper/hashPassword.js
