+++ backend/src/helper/validateUser.js
