+++ backend/src/middleware/unknownEndpoint.js
@@ -1,4 +1,4 @@
-const { ErrorHandler } = require("../helpers/error");
+const { ErrorHandler } = require("../helper/error");
 
 const unknownEndpoint = (req, res) => {
     throw new ErrorHandler(401, "Unknown Endpoint");
