+++ backend/src/route/index.js
@@ -1,12 +1,11 @@
 const router = require("express").Router();
 
-const users = require("./users");
+const user = require("./user");
 const auth = require("./auth");
 const sample = require("./sample")
 
 router.use("/auth", auth);
-router.use("/users", users);
+router.use("/user", user);
 router.use("/sample", sample);
-
 
 module.exports = router;
